fix(primitives): extract PreCommitSectorInfo
