+++ pallets/storage-provider/Cargo.toml
@@ -38,6 +38,7 @@ env_logger = { workspace = true }
 multihash-codetable = { workspace = true, features = ["blake2b"] }
 pallet-balances = { workspace = true, default-features = false }
 pallet-market = { workspace = true, default-features = false }
+primitives = { workspace = true, features = ["builder"] }
 rstest = { workspace = true }
 sp-io = { workspace = true }
