+++ pallets/storage-provider/src/tests/prove_commit_sectors.rs
@@ -1,6 +1,6 @@
 use frame_support::{assert_noop, assert_ok, pallet_prelude::*};
 use frame_system::pallet_prelude::BlockNumberFor;
-use primitives::MAX_SECTORS_PER_CALL;
+use primitives::{sector::SectorPreCommitInfo, MAX_SECTORS_PER_CALL};
 use sp_core::bounded_vec;
 
 use super::{new_test_ext, MaxProveCommitDuration};
@@ -8,7 +8,7 @@ use crate::{
     deadline::deadline_is_mutable,
     error::GeneralPalletError,
     pallet::{Error, Event, StorageProviders},
-    sector::{ProveCommitResult, ProveCommitSector, SectorPreCommitInfo},
+    sector::{ProveCommitResult, ProveCommitSector},
     tests::{
         account, events, publish_deals, register_storage_provider, run_to_block, Balances,
         RuntimeEvent, RuntimeOrigin, SectorPreCommitInfoBuilder, StorageProvider, System, Test,
